change url
I change url from '/info?id=1' to '/1'. Because appear a bug when you use '?' on url.

+++ src/main/java/com/example/medimateserver/controller/api/CouponDetailController.java
@@ -7,7 +7,6 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.math.BigInteger;
 import java.util.List;
 
 @RestController
@@ -23,7 +22,7 @@ public class CouponDetailController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<CouponDetail> getCouponDetailById(@PathVariable BigInteger id) {
+    public ResponseEntity<CouponDetail> getCouponDetailById(@PathVariable Integer id) {
         CouponDetail CouponDetail = couponDetailService.findById(id);
         if (CouponDetail != null) {
             return ResponseEntity.ok(CouponDetail);
