+++ src/main/java/com/example/medimateserver/dto/CategoryDto.java
@@ -2,10 +2,9 @@ package com.example.medimateserver.dto;
 
 import jakarta.persistence.Column;
 
-import java.math.BigInteger;
 
 public class CategoryDto {
-    private BigInteger id;
+    private Integer id;
     private String name;
     private Integer status;
 }
