+++ src/main/java/com/example/medimateserver/dto/CouponDetailDto.java
@@ -2,13 +2,12 @@ package com.example.medimateserver.dto;
 
 import jakarta.persistence.Column;
 
-import java.math.BigInteger;
 import java.util.Date;
 
 public class CouponDetailDto {
-    private BigInteger id;
-    private BigInteger idUser;
-    private BigInteger idPromoRule;
+    private Integer id;
+    private Integer idUser;
+    private Integer idPromoRule;
     private String code;
     private Date timeStart;
     private Date timeEnd;
