+++ src/main/java/com/example/medimateserver/dto/CouponDto.java
@@ -2,13 +2,12 @@ package com.example.medimateserver.dto;
 
 import jakarta.persistence.Column;
 
-import java.math.BigInteger;
 import java.util.Date;
 
 public class CouponDto {
 
-    private BigInteger id;
-    private BigInteger necessaryPoints;
+    private Integer id;
+    private Integer necessaryPoints;
     private Integer percentDecrease;
     private Integer status;
 }
