+++ src/main/java/com/example/medimateserver/dto/OrderDetailDto.java
@@ -2,10 +2,9 @@ package com.example.medimateserver.dto;
 
 import jakarta.persistence.Column;
 
-import java.math.BigInteger;
 
 public class OrderDetailDto {
-    private BigInteger idOrder;
-    private BigInteger idProduct;
+    private Integer idOrder;
+    private Integer idProduct;
     private Integer quantity;
 }
