+++ src/main/java/com/example/medimateserver/dto/OrderDto.java
@@ -2,12 +2,10 @@ package com.example.medimateserver.dto;
 
 import jakarta.persistence.Column;
 
-import java.math.BigInteger;
-
 public class OrderDto {
-    private BigInteger id;
-    private BigInteger idUser;
-    private BigInteger idPromo;
+    private Integer id;
+    private Integer idUser;
+    private Integer idPromo;
     private Integer paymentMethod;
     private Integer status;
 }
