+++ src/main/java/com/example/medimateserver/dto/ProductDto.java
@@ -1,10 +1,9 @@
 package com.example.medimateserver.dto;
 
-import java.math.BigInteger;
 
 public class ProductDto {
-    private BigInteger id;
-    private BigInteger idCategory;
+    private Integer id;
+    private Integer idCategory;
     private String name;
     private String description;
     private Integer percent;
