+++ src/main/java/com/example/medimateserver/model/Order.java
@@ -4,8 +4,6 @@ import jakarta.persistence.*;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigInteger;
-
 @Entity
 @Table(name = "order")
 @Data
@@ -13,11 +11,11 @@ import java.math.BigInteger;
 public class Order {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private BigInteger id;
+    private Integer id;
     @Column(name = "id_user")
-    private BigInteger idUser;
+    private Integer idUser;
     @Column(name = "id_promo")
-    private BigInteger idPromo;
+    private Integer idPromo;
     @Column(name = "payment_method")
     private Integer paymentMethod;
     @Column(name = "status")
