+++ src/main/java/com/example/medimateserver/service/CategoryService.java
@@ -3,19 +3,18 @@ package com.example.medimateserver.service;
 import com.example.medimateserver.model.Category;
 import org.springframework.stereotype.Service;
 
-import java.math.BigInteger;
 import java.util.List;
 @Service
 public interface CategoryService {
 
     List<Category> findAll();
 
-    Category findById(BigInteger id);
+    Category findById(Integer id);
     Category findByEmail(String email);
 
     Category save(Category category);
 
-    Category update(BigInteger id, Category category);
+    Category update(Integer id, Category category);
 
-    void deleteById(BigInteger id);
+    void deleteById(Integer id);
 }
