+++ src/main/java/com/example/medimateserver/service/CouponService.java
@@ -4,7 +4,6 @@ import com.example.medimateserver.model.Coupon;
 import com.example.medimateserver.model.Coupon;
 import org.springframework.stereotype.Service;
 
-import java.math.BigInteger;
 import java.util.List;
 
 @Service
@@ -12,13 +11,13 @@ public interface CouponService {
 
     List<Coupon> findAll();
 
-    Coupon findById(BigInteger id);
+    Coupon findById(Integer id);
 
     Coupon save(Coupon Coupon);
 
-    Coupon update(BigInteger id, Coupon coupon);
+    Coupon update(Integer id, Coupon coupon);
 
-    void deleteById(BigInteger id);
+    void deleteById(Integer id);
 
     Coupon findByCode(String code);
 }
