+++ src/main/java/com/example/medimateserver/service/OrderService.java
@@ -3,19 +3,18 @@ package com.example.medimateserver.service;
 import com.example.medimateserver.model.Order;
 import org.springframework.stereotype.Service;
 
-import java.math.BigInteger;
 import java.util.List;
 @Service
 public interface OrderService {
 
     List<Order> findAll();
 
-    Order findById(BigInteger id);
+    Order findById(Integer id);
     Order findByEmail(String email);
 
     Order save(Order order);
 
-    Order update(BigInteger id, Order order);
+    Order update(Integer id, Order order);
 
-    void deleteById(BigInteger id);
+    void deleteById(Integer id);
 }
